requirements.txt
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
ConfigParser==7.2.0
cryptography==45.0.5
curio==1.6
cython==3.1.2
docutils==0.21.2
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi_cli==0.0.8
filelock==3.18.0
h2==4.2.0
HTMLParser==0.0.2
httpx==0.28.1
hypothesis==6.136.4
ipython==9.4.0
ipywidgets==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.6
jnius==1.1.0
js==1.0
keyring==25.6.0
multipart==1.2.1
orjson==3.11.1
outcome==1.3.0.post0
protobuf==6.31.1
pyodide==0.0.2
pyOpenSSL==25.1.0
pytest==8.4.1
python-dotenv==1.1.1
python_multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
simplejson==3.20.1
Sphinx==8.2.3
thread==2.0.5
toml==0.10.2
trio==0.30.0
ujson==5.10.0
urllib3_secure_extra==0.1.0
uvloop==0.21.0
xmlrpclib==1.0.1
zstandard==0.23.0
uvicorn==0.27.0